requirements.txt
requests>=2.28.0
APScheduler>=3.10.0
crewai>=0.80.0
crewai-tools>=0.14.0
crewai[litellm]>=0.80.0
duckduckgo-search>=4.0.0
python-dotenv>=1.0.0
feedparser>=6.0.0
google-api-python-client>=2.0.0
markdown>=3.0.0
